fix: Resolve MSVC warnings and CI build failures on Windows

+++ src/vm/Chunk.h
@@ -157,7 +157,7 @@ public:
             }
         }
         constants.push_back(value);
-        return constants.size() - 1;
+        return static_cast<int>(constants.size() - 1);
     }
 };
